# docs/swagger.yaml
definitions:
  base.Output:
    properties:
      code:
        description: 狀態碼
        example: 9000
        type: integer
      msg:
        description: 訊息
        example: message..
        type: string
    type: object
...
  course.APIGetCMSCourseOutput:
    properties:
      code:
        description: 狀態碼
        example: 9000
        type: integer
      data:
        $ref: '#/definitions/course.APIGetCMSCourseData'
      msg:
        description: 訊息
        example: message..
        type: string
    type: object
  course.APIGetCMSCoursesOutput:
    properties:
      code:
        description: 狀態碼
        example: 9000
        type: integer
      data:
        items:
          properties:
            course_status:
              description: 課表狀態 (1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
              example: 3
              type: integer
            create_at:
              description: 創建時間
              example: "2022-06-12 00:00:00"
              type: string
            id:
              description: 課表 id
              example: 2
              type: integer
            name:
              description: 課表名稱
              example: 增肌課表
              type: string
            sale_item:
              properties:
                id:
                  description: 銷售id
                  example: 1
                  type: integer
                name:
                  description: 銷售名稱
                  example: '銅級課表 '
                  type: string
                product_label:
                  properties:
                    id:
                      description: 產品標籤id
                      example: 1
                      type: integer
                    product_id:
                      description: 產品id
                      example: com.fitness.gold_member_month
                      type: string
                    twd:
                      description: 台幣價格
                      example: 500
                      type: integer
                  type: object
              type: object
            sale_type:
              description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
              example: 3
              type: integer
            schedule_type:
              description: 排課類別(1:單一訓練/2:多項計畫)
              example: 2
              type: integer
            trainer:
              properties:
                nickname:
                  description: 教練暱稱
                  example: Henry
                  type: string
                user_id:
                  description: 用戶id
                  example: 10001
                  type: integer
              type: object
          type: object
        type: array
      msg:
        description: 訊息
        example: message..
        type: string
      paging:
        $ref: '#/definitions/paging.Output'
    type: object
...
  dto.CourseProduct:
    properties:
      allow_access:
        description: 是否允許訪問此課表(0:否/1:是)
        example: 0
        type: integer
      body_target:
        description: 體態目標(1:比基尼身材/2:翹臀/3:健力/4:健美/5:腹肌/6:馬甲線/7:其他)
        example: 4,5
        type: string
      category:
        description: 課表類別(1:有氧心肺訓練/2:間歇肌力訓練/3:重量訓練/4:阻力訓練/5:徒手訓練/6:其他)
        example: 3
        type: integer
      course_status:
        description: 課表狀態(1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
        example: 1
        type: integer
      cover:
        description: 課表封面
        example: d2w3e15d3awe.jpg
        type: string
      equipment:
        description: 所需器材(1:無需任何器材/2:啞鈴/3:槓鈴/4:固定式器材/5:彈力繩/6:壺鈴/7:訓練椅/8:瑜珈墊/9:其他)
        example: 2,3,6
        type: string
      favorite:
        description: 是否收藏(0:否/1:是)
        example: 1
        type: integer
      food:
        description: 飲食建議
        example: 佛系飲食
        type: string
      id:
        description: 課表 id
        example: 2
        type: integer
      intro:
        description: 課表介紹
        example: 佛系課表
        type: string
      level:
        description: 強度(1:初級/2:中級/3:中高級/4:高級)
        example: 3
        type: integer
      name:
        description: 課表名稱
        example: Henry課表
        type: string
      notice:
        description: 注意事項
        example: 小心不要受傷
        type: string
      place:
        description: 適合場地(1:健身房/2:居家/3:空地/4:戶外/5:其他)
        example: 1,2
        type: string
      plan_count:
        description: 計畫總數
        example: 2
        type: integer
      review:
        $ref: '#/definitions/dto.ReviewStatistic'
        description: 評分統計
      sale:
        $ref: '#/definitions/dto.SaleItem'
        description: 銷售項目
      sale_type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 2
        type: integer
      schedule_type:
        description: 排課類別(1:單一訓練/2:多項計畫)
        example: 2
        type: integer
      suit:
        description: 適用對象(1:女性/2:男性/3:初學者/4:進階者/5:專業/6:長輩/7:運動員/8:孕婦/9:產後/10:其他)
        example: 2,5,7
        type: string
      train_target:
        description: 訓練目的(1:減脂/2:增肌/3:維持健康/4:鐵人三項/5:其他)
        example: 2,3,4
        type: string
      trainer:
        $ref: '#/definitions/dto.TrainerSummary'
        description: 教練簡介
      workout_count:
        description: 訓練總數
        example: 10
        type: integer
    type: object
  dto.CourseProductStructure:
    properties:
      allow_access:
        description: 是否允許訪問此課表(0:否/1:是)
        example: 0
        type: integer
      body_target:
        description: 體態目標(1:比基尼身材/2:翹臀/3:健力/4:健美/5:腹肌/6:馬甲線/7:其他)
        example: 4,5
        type: string
      category:
        description: 課表類別(1:有氧心肺訓練/2:間歇肌力訓練/3:重量訓練/4:阻力訓練/5:徒手訓練/6:其他)
        example: 3
        type: integer
      course_status:
        description: 課表狀態(1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
        example: 1
        type: integer
      cover:
        description: 課表封面
        example: d2w3e15d3awe.jpg
        type: string
      equipment:
        description: 所需器材(1:無需任何器材/2:啞鈴/3:槓鈴/4:固定式器材/5:彈力繩/6:壺鈴/7:訓練椅/8:瑜珈墊/9:其他)
        example: 2,3,6
        type: string
      favorite:
        description: 是否收藏(0:否/1:是)
        example: 1
        type: integer
      food:
        description: 飲食建議
        example: 佛系飲食
        type: string
      id:
        description: 課表 id
        example: 2
        type: integer
      intro:
        description: 課表介紹
        example: 佛系課表
        type: string
      level:
        description: 強度(1:初級/2:中級/3:中高級/4:高級)
        example: 3
        type: integer
      name:
        description: 課表名稱
        example: Henry課表
        type: string
      notice:
        description: 注意事項
        example: 小心不要受傷
        type: string
      place:
        description: 適合場地(1:健身房/2:居家/3:空地/4:戶外/5:其他)
        example: 1,2
        type: string
      plan_count:
        description: 計畫總數
        example: 2
        type: integer
      plans:
        description: 計畫列表
        items:
          $ref: '#/definitions/dto.PlanStructure'
        type: array
      review:
        $ref: '#/definitions/dto.ReviewStatistic'
        description: 評分統計
      sale:
        $ref: '#/definitions/dto.SaleItem'
        description: 銷售項目
      sale_type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 2
        type: integer
      schedule_type:
        description: 排課類別(1:單一訓練/2:多項計畫)
        example: 2
        type: integer
      suit:
        description: 適用對象(1:女性/2:男性/3:初學者/4:進階者/5:專業/6:長輩/7:運動員/8:孕婦/9:產後/10:其他)
        example: 2,5,7
        type: string
      train_target:
        description: 訓練目的(1:減脂/2:增肌/3:維持健康/4:鐵人三項/5:其他)
        example: 2,3,4
        type: string
      trainer:
        $ref: '#/definitions/dto.TrainerSummary'
        description: 教練簡介
      workout_count:
        description: 訓練總數
        example: 10
        type: integer
    type: object
  dto.CourseProductSummary:
    properties:
      category:
        description: 課表類別(1:有氧心肺訓練/2:間歇肌力訓練/3:重量訓練/4:阻力訓練/5:徒手訓練/6:其他)
        example: 3
        type: integer
      course_status:
        description: 課表狀態 (1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
        example: 1
        type: integer
      cover:
        description: 課表封面
        example: d2w3e15d3awe.jpg
        type: string
      id:
        description: 課表 id
        example: 2
        type: integer
      level:
        description: 強度(1:初級/2:中級/3:中高級/4:高級)
        example: 3
        type: integer
      name:
        description: 課表名稱
        example: Henry課表
        type: string
      plan_count:
        description: 計畫總數
        example: 2
        type: integer
      review:
        $ref: '#/definitions/dto.ReviewStatisticSummary'
        description: 評分統計
      sale:
        $ref: '#/definitions/dto.SaleItem'
        description: 銷售項目
      sale_type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 2
        type: integer
      schedule_type:
        description: 排課類別(1:單一訓練/2:多項計畫)
        example: 2
        type: integer
      trainer:
        $ref: '#/definitions/dto.TrainerSummary'
        description: 教練簡介
      workout_count:
        description: 訓練總數
        example: 10
        type: integer
    type: object
  dto.CourseStatistic:
    properties:
      category:
        description: 課表類別(1:有氧心肺訓練/2:間歇肌力訓練/3:重量訓練/4:阻力訓練/5:徒手訓練/6:其他)
        example: 3
        type: integer
      course_status:
        description: 課表狀態 (1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
        example: 1
        type: integer
      course_usage_statistic:
        $ref: '#/definitions/dto.CourseUsageStatistic'
        description: 課表使用統計
      cover:
        description: 課表封面
        example: d2w3e15d3awe.jpg
        type: string
      id:
        description: 課表 id
        example: 2
        type: integer
      level:
        description: 強度(1:初級/2:中級/3:中高級/4:高級)
        example: 3
        type: integer
      name:
        description: 課表名稱
        example: Henry課表
        type: string
      plan_count:
        description: 計畫總數
        example: 2
        type: integer
      review:
        $ref: '#/definitions/dto.ReviewStatisticSummary'
        description: 評分統計
      sale_type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 2
        type: integer
      schedule_type:
        description: 排課類別(1:單一訓練/2:多項計畫)
        example: 2
        type: integer
      user_id:
        description: 教練 id
        example: 10001
        type: integer
      workout_count:
        description: 訓練總數
        example: 10
        type: integer
    type: object
  dto.CourseStatisticSummary:
    properties:
      category:
        description: 課表類別(1:有氧心肺訓練/2:間歇肌力訓練/3:重量訓練/4:阻力訓練/5:徒手訓練/6:其他)
        example: 3
        type: integer
      course_status:
        description: 課表狀態 (1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
        example: 1
        type: integer
      course_usage_statistic:
        $ref: '#/definitions/dto.CourseUsageStatisticSummary'
        description: 課表使用統計
      cover:
        description: 課表封面
        example: d2w3e15d3awe.jpg
        type: string
      id:
        description: 課表 id
        example: 1
        type: integer
      level:
        description: 強度(1:初級/2:中級/3:中高級/4:高級)
        example: 3
        type: integer
      name:
        description: 課表名稱
        example: Henry課表
        type: string
      plan_count:
        description: 計畫總數
        example: 100
        type: integer
      sale_type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 1
        type: integer
      schedule_type:
        description: 排課類別(1:單一訓練/2:多項計畫)
        example: 2
        type: integer
      user_id:
        description: 教練 id
        example: 10001
        type: integer
      workout_count:
        description: 訓練總數
        example: 10
        type: integer
    type: object
...
  dto.Plan:
    properties:
      id:
        description: 計畫id
        example: 1
        type: integer
      name:
        description: 計畫名稱
        example: 第一週增肌計畫
        type: string
      workout_count:
        description: 包含訓練數量
        example: 10
        type: integer
    type: object
...
  dto.ReviewStatistic:
    properties:
      amount:
        description: 評分筆數
        example: 450
        type: integer
      five_total:
        description: 五分總筆數
        example: 100
        type: integer
      four_total:
        description: 四分總筆數
        example: 100
        type: integer
      one_total:
        description: 一分總筆數
        example: 50
        type: integer
      score_total:
        description: 評分累積
        example: 1000
        type: integer
      three_total:
        description: 三分總筆數
        example: 150
        type: integer
      two_total:
        description: 二分總筆數
        example: 100
        type: integer
      update_at:
        description: 更新時間
        example: "100"
        type: string
    type: object
...
  dto.SaleItem:
    properties:
      id:
        description: 銷售id
        example: 1
        type: integer
      name:
        description: 銷售名稱
        example: 銅級課表
        type: string
      product_id:
        description: 產品ID
        example: com.fitness.xxx
        type: string
      twd:
        description: 台幣價格
        example: 330
        type: integer
      type:
        description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        example: 3
        type: integer
    type: object
...
  dto.TrainerSummary:
    properties:
      avatar:
        description: 教練大頭照
        example: d2w3e15d3awe.jpg
        type: string
      nickname:
        description: 教練暱稱
        example: Henry教練
        type: string
      skill:
        description: 專長
        example: d2w3e15d3awe.jpg
        type: string
      user_id:
        description: 關聯的用戶id
        example: 10001
        type: integer
    type: object
...
  model.Data:
    type: object
  model.ErrorResult:
    properties:
      code:
        description: 錯誤碼
        example: 9000
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 錯誤訊息
        example: system error!
        type: string
    type: object
...
  model.SuccessPagingResult:
    properties:
      code:
        description: 狀態碼
        example: 0
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 成功訊息
        example: success!
        type: string
      paging:
        $ref: '#/definitions/model.Paging'
        description: 分頁資訊
    type: object
  model.SuccessResult:
    properties:
      code:
        description: 狀態碼
        example: 0
        type: integer
      data:
        $ref: '#/definitions/model.Data'
        description: 回傳資料
      msg:
        description: 成功訊息
        example: success!
        type: string
    type: object
...
  paging.Output:
    properties:
      page:
        description: 當前頁數
        example: 1
        type: integer
      size:
        description: 一頁筆數
        example: 5
        type: integer
      total_count:
        description: 總筆數
        example: 100
        type: integer
      total_page:
        description: 總頁數
        example: 10
        type: integer
    required:
    - page
    - size
    type: object
...
info:
  contact: {}
  description: 健身平台 api
  title: fitness api
paths:
  /cms/course/{id}:
    get:
      consumes:
      - application/json
      description: 獲取課表詳細
      parameters:
      - description: 課表ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            $ref: '#/definitions/course.APIGetCMSCourseOutput'
        "400":
          description: 失敗!
          schema:
            $ref: '#/definitions/base.Output'
      security:
      - fitness_token: []
      summary: 獲取課表詳細
      tags:
      - CMS/Course_v2
  /cms/courses:
    get:
      consumes:
      - application/json
      description: 獲取課表列表
      parameters:
      - description: 課表ID
        in: query
        name: id
        type: integer
      - description: 課表名稱
        in: query
        name: name
        type: string
      - description: 課表狀態 (1:準備中/2:審核中/3:銷售中/4:退審/5:下架)
        in: query
        name: course_status
        type: integer
      - description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)
        in: query
        name: sale_type
        type: integer
      - description: 排序欄位 (create_at:創建時間)
        in: query
        name: order_field
        required: true
        type: string
      - description: 排序類型 (ASC:由低到高/DESC:由高到低)
        in: query
        name: order_type
        required: true
        type: string
      - description: 頁數(從第一頁開始)
        in: query
        name: page
        required: true
        type: integer
      - description: 筆數
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功!
          schema:
            $ref: '#/definitions/course.APIGetCMSCoursesOutput'
        "400":
          description: 失敗!
          schema:
            $ref: '#/definitions/base.Output'
      security:
      - fitness_token: []
      summary: 獲取課表列表
      tags:
      - CMS/Course_v2
...
  /v1/course_product/{course_id}:
    get:
      consumes:
      - application/json
      description: 獲取課表產品詳細
      parameters:
      - description: 課表id
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.CourseProduct'
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取課表產品詳細
      tags:
      - Explore_v1
  /v1/course_product/{course_id}/plans:
    get:
      consumes:
      - application/json
      description: 獲取課表產品計畫列表
      parameters:
      - description: 課表id
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Plan'
                  type: array
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取課表產品計畫列表
      tags:
      - Explore_v1
...
  /v1/course_product_structure/{course_id}:
    get:
      consumes:
      - application/json
      description: 只限單一訓練的課表使用
      parameters:
      - description: 課表id
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.CourseProductStructure'
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取課表結構(只限單一訓練課表)
      tags:
      - Explore_v1
  /v1/course_products:
    get:
      consumes:
      - application/json
      description: 搜尋課表產品列表
      parameters:
      - description: 課表名稱(1~20字元)
        in: query
        name: name
        type: string
      - description: 排序類型(latest:最新/popular:熱門)
        in: query
        name: order_type
        type: string
      - description: 評價(1~5分)-單選
        in: query
        name: score
        type: integer
      - description: 強度(1:初級/2:中級/3:中高級/4:高級)-複選
        in: query
        name: level
        type: integer
      - description: 課表類別(1:有氧心肺訓練/2:間歇肌力訓練/3:重量訓練/4:阻力訓練/5:徒手訓練/6:其他)-複選
        in: query
        name: category
        type: integer
      - description: 適用對象(1:女性/2:男性/3:初學者/4:進階者/5:專業/6:長輩/7:運動員/8:孕婦/9:產後/10:其他)-複選
        in: query
        name: suit
        type: integer
      - description: 所需器材(1:無需任何器材/2:啞鈴/3:槓鈴/4:固定式器材/5:彈力繩/6:壺鈴/7:訓練椅/8:瑜珈墊/9:其他)-複選
        in: query
        name: equipment
        type: integer
      - description: 適合場地(1:健身房/2:居家/3:空地/4:戶外/5:其他)-複選
        in: query
        name: place
        type: integer
      - description: 訓練目的(1:減脂/2:增肌/3:維持健康/4:鐵人三項/5:其他)-複選
        in: query
        name: train_target
        type: integer
      - description: 體態目標(1:比基尼身材/2:翹臀/3:健力/4:健美/5:腹肌/6:馬甲線/7:其他)-複選
        in: query
        name: body_target
        type: integer
      - description: 銷售類型(1:免費課表/2:訂閱課表/3:付費課表)-複選
        in: query
        name: sale_type
        type: integer
      - description: 教練性別(m:男性/f:女性)-複選
        in: query
        name: trainer_sex
        type: string
      - description: 教練專長(1:功能性訓練/2:減脂/3:增肌/4:健美規劃/5:運動項目訓練/6:TRX/7:重量訓練/8:筋膜放鬆/9:瑜珈/10:體態雕塑/11:減重/12:心肺訓練/13:肌力訓練/14:其他)
        in: query
        name: trainer_skill
        type: integer
      - description: 頁數(從第一頁開始)
        in: query
        name: page
        required: true
        type: integer
      - description: 筆數
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessPagingResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CourseProductSummary'
                  type: array
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 搜尋課表產品列表
      tags:
      - Explore_v1
  /v1/course_statistic/{course_id}:
    get:
      consumes:
      - application/json
      description: 獲取個人課表數據詳細
      parameters:
      - description: 課表id
        in: path
        name: course_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.CourseStatistic'
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取個人課表數據詳細
      tags:
      - Statistic_v1
  /v1/course_statistics:
    get:
      consumes:
      - application/json
      description: 獲取個人課表數據統計列表
      parameters:
      - description: 頁數(從第一頁開始)
        in: query
        name: page
        required: true
        type: integer
      - description: 筆數
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 獲取成功!
          schema:
            allOf:
            - $ref: '#/definitions/model.SuccessPagingResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CourseStatisticSummary'
                  type: array
              type: object
        "400":
          description: 獲取失敗
          schema:
            $ref: '#/definitions/model.ErrorResult'
      security:
      - fitness_token: []
      summary: 獲取個人課表數據統計列表
      tags:
      - Statistic_v1
...
securityDefinitions:
  fitness_token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"